factor out post types
